data requst method and put data logic in hooks
